+++ source/solvers/tracer.cc
@@ -35,14 +35,15 @@ template <int dim>
 void
 Tracer<dim>::setup_assemblers()
 {
-  AssertThrow(
-    is_sdirk(this->simulation_control->get_assembly_method()) == false,
-    ExcMessage("The SDIRK scheme is not yet supported for this physics"));
+  AssertThrow(time_stepping_is_sdirk(
+                this->simulation_control->get_assembly_method()) == false,
+              ExcMessage(
+                "The SDIRK scheme is not yet supported for this physics"));
 
   this->assemblers.clear();
 
   // Time-stepping schemes
-  if (is_bdf(this->simulation_control->get_assembly_method()))
+  if (time_stepping_is_bdf(this->simulation_control->get_assembly_method()))
     {
       this->assemblers.emplace_back(
         std::make_shared<TracerAssemblerBDF<dim>>(this->simulation_control));