+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the Lethe project will be documented in this file.
 The changelog for the previous releases of Lethe are located in the release_notes folder.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+### [Master] - 2025-11-25
+
+### Added
+
+- MINOR Add the SDIRK time integration method to the lethe-fluid-matrix-free solver. The SDIRK scheme is now available for both MatrixBased and MatrixFree solvers. However, it does not support multiphysics components at the moment. This will be added in a near future. [#1616](https://github.com/chaos-polymtl/lethe/pull/1616/)
+
 
 ### [Master] - 2025-11-23
 
@@ -42,9 +48,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### [Master] - 2025-11-14
 
-- MINOR Remove the deprecated argument "renewed_matrix" of the solve_linear_system definition. 
+### Removed
 
-### [Master] - 2025-11-14
+- MINOR Remove the deprecated argument "renewed_matrix" of the solve_linear_system definition. 
 
 ### Fixed